Separated analytics and stats backend
